refactor: move around some configuration code again
